requirements.txt
asgiref==3.9.1
blinker==1.9.0
click==8.3.0
colorama==0.4.6
Django==5.2.6
django-cors-headers==4.8.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
Flask==3.1.2
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==25.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
sqlparse==0.5.3
tzdata==2025.2
Werkzeug==3.1.3
